+++ picasso_workflow/confluence.py
@@ -4976,6 +4976,7 @@ class ConfluenceReporter(AbstractModuleCollection):
         for _k in (
             "fp_pattern_phasespace",
             "fp_pattern_fit_space",
+            "fp_pattern_pairscore_space",
             "fp_pattern_site_nlocs_hist",
             "fp_pattern_feature_space",
             "fp_pattern_pairdist",
@@ -5047,6 +5048,7 @@ class ConfluenceReporter(AbstractModuleCollection):
             for _key in (
                 "fp_pattern_phasespace",
                 "fp_pattern_fit_space",
+                "fp_pattern_pairscore_space",
                 "fp_pattern_site_nlocs_hist",
                 "fp_pattern_feature_space",
                 "fp_pattern_pairdist",